surreal-prompt-bot/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
huggingface_hub>=0.20.0
slack-sdk>=3.27.0
pyyaml>=6.0.0
pytest>=8.0.0
